Make Decoded struct public, refactor
