+++ src/encoding.rs
@@ -7,19 +7,7 @@
 // except according to those terms.
 
 use crate::context::Context;
-use crate::decoded::Decoded;
-use crate::error::Error;
-use crate::result::Result;
-use crate::variant::Variant;
-use crate::version::Version;
 use base64;
-
-/// Structure containing the options.
-struct Options {
-    mem_cost: u32,
-    time_cost: u32,
-    parallelism: u32,
-}
 
 /// Gets the base64 encoded length of a byte slice with the specified length.
 pub fn base64_len(length: u32) -> u32 {
@@ -28,105 +16,6 @@ pub fn base64_len(length: u32) -> u32 {
         2 => olen + 3,
         1 => olen + 2,
         _ => olen,
-    }
-}
-
-/// Attempts to decode the encoded string slice.
-pub fn decode_string(encoded: &str) -> Result<Decoded> {
-    let items: Vec<&str> = encoded.split('$').collect();
-    if items.len() == 6 {
-        decode_empty(items[0])?;
-        let variant = decode_variant(items[1])?;
-        let version = decode_version(items[2])?;
-        let options = decode_options(items[3])?;
-        let salt = base64::decode(items[4])?;
-        let hash = base64::decode(items[5])?;
-
-        Ok(Decoded {
-            variant,
-            version,
-            mem_cost: options.mem_cost,
-            time_cost: options.time_cost,
-            parallelism: options.parallelism,
-            salt,
-            hash,
-        })
-    } else if items.len() == 5 {
-        decode_empty(items[0])?;
-        let variant = decode_variant(items[1])?;
-        let options = decode_options(items[2])?;
-        let salt = base64::decode(items[3])?;
-        let hash = base64::decode(items[4])?;
-
-        Ok(Decoded {
-            variant,
-            version: Version::Version10,
-            mem_cost: options.mem_cost,
-            time_cost: options.time_cost,
-            parallelism: options.parallelism,
-            salt,
-            hash,
-        })
-    } else {
-        Err(Error::DecodingFail)
-    }
-}
-
-fn decode_empty(str: &str) -> Result<()> {
-    if str == "" {
-        Ok(())
-    } else {
-        Err(Error::DecodingFail)
-    }
-}
-
-fn decode_options(str: &str) -> Result<Options> {
-    let items: Vec<&str> = str.split(',').collect();
-    if items.len() == 3 {
-        Ok(Options {
-            mem_cost: decode_option(items[0], "m")?,
-            time_cost: decode_option(items[1], "t")?,
-            parallelism: decode_option(items[2], "p")?,
-        })
-    } else {
-        Err(Error::DecodingFail)
-    }
-}
-
-fn decode_option(str: &str, name: &str) -> Result<u32> {
-    let items: Vec<&str> = str.split('=').collect();
-    if items.len() == 2 {
-        if items[0] == name {
-            decode_u32(items[1])
-        } else {
-            Err(Error::DecodingFail)
-        }
-    } else {
-        Err(Error::DecodingFail)
-    }
-}
-
-fn decode_u32(str: &str) -> Result<u32> {
-    match str.parse() {
-        Ok(i) => Ok(i),
-        Err(_) => Err(Error::DecodingFail),
-    }
-}
-
-fn decode_variant(str: &str) -> Result<Variant> {
-    Variant::from_str(str)
-}
-
-fn decode_version(str: &str) -> Result<Version> {
-    let items: Vec<&str> = str.split('=').collect();
-    if items.len() == 2 {
-        if items[0] == "v" {
-            Version::from_str(items[1])
-        } else {
-            Err(Error::DecodingFail)
-        }
-    } else {
-        Err(Error::DecodingFail)
     }
 }
 
@@ -165,7 +54,7 @@ mod tests {
     use crate::decoded::Decoded;
     #[cfg(feature = "crossbeam-utils")]
     use crate::encoding::encode_string;
-    use crate::encoding::{base64_len, decode_string, num_len};
+    use crate::encoding::{base64_len, num_len};
     use crate::error::Error;
     #[cfg(feature = "crossbeam-utils")]
     use crate::thread_mode::ThreadMode;
@@ -190,177 +79,6 @@ mod tests {
             let actual = base64_len(len);
             assert_eq!(actual, expected);
         }
-    }
-
-    #[test]
-    fn decode_string_with_version10_returns_correct_result() {
-        let encoded = "$argon2i$v=16$m=4096,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let expected = Decoded {
-            variant: Variant::Argon2i,
-            version: Version::Version10,
-            mem_cost: 4096,
-            time_cost: 3,
-            parallelism: 1,
-            salt: b"salt1234".to_vec(),
-            hash: b"12345678901234567890123456789012".to_vec(),
-        };
-        let actual = decode_string(encoded).unwrap();
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn decode_string_with_version13_returns_correct_result() {
-        let encoded = "$argon2i$v=19$m=4096,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let expected = Decoded {
-            variant: Variant::Argon2i,
-            version: Version::Version13,
-            mem_cost: 4096,
-            time_cost: 3,
-            parallelism: 1,
-            salt: b"salt1234".to_vec(),
-            hash: b"12345678901234567890123456789012".to_vec(),
-        };
-        let actual = decode_string(encoded).unwrap();
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn decode_string_without_version_returns_correct_result() {
-        let encoded = "$argon2i$m=4096,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let expected = Decoded {
-            variant: Variant::Argon2i,
-            version: Version::Version10,
-            mem_cost: 4096,
-            time_cost: 3,
-            parallelism: 1,
-            salt: b"salt1234".to_vec(),
-            hash: b"12345678901234567890123456789012".to_vec(),
-        };
-        let actual = decode_string(encoded).unwrap();
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn decode_string_without_variant_returns_error_result() {
-        let encoded = "$m=4096,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_empty_variant_returns_error_result() {
-        let encoded = "$$m=4096,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_invalid_variant_returns_error_result() {
-        let encoded = "$argon$m=4096,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_without_mem_cost_returns_error_result() {
-        let encoded = "$argon2i$t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_empty_mem_cost_returns_error_result() {
-        let encoded = "$argon2i$m=,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_non_numeric_mem_cost_returns_error_result() {
-        let encoded = "$argon2i$m=a,t=3,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_without_time_cost_returns_error_result() {
-        let encoded = "$argon2i$m=4096,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_empty_time_cost_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_non_numeric_time_cost_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=a,p=1\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_without_parallelism_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=3\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_empty_parallelism_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=3,p=\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_non_numeric_parallelism_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=3,p=a\
-                       $c2FsdDEyMzQ=$MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_without_salt_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=3,p=1\
-                       $MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTI=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_without_hash_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=3,p=a\
-                       $c2FsdDEyMzQ=";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
-    }
-
-    #[test]
-    fn decode_string_with_empty_hash_returns_error_result() {
-        let encoded = "$argon2i$m=4096,t=3,p=a\
-                       $c2FsdDEyMzQ=$";
-        let result = decode_string(encoded);
-        assert_eq!(result, Err(Error::DecodingFail));
     }
 
     #[cfg(feature = "crossbeam-utils")]
